More ANSI changes
git-svn-id: https://svn.so-much-stuff.com/svn/trunk/cvs/trunk@1525 93276802-5317-4dee-a4ef-7ae40ebda4b9

+++ games.d/nethack3.0.d/src/decl.c
@@ -5,7 +5,7 @@
 #include	"hack.h"
 
 int multi = 0;
-int warnlevel = 0;		/* used by movemon and dochugw */
+uchar warnlevel = 0;		/* used by movemon and dochugw */
 int nroom = 0;
 int occtime = 0;
